+++ src/main/kotlin/com/github/victorrentea/livecoding/settings/AppSettingsComponent.kt
@@ -13,15 +13,15 @@ class AppSettingsComponent {
     fun getPreferredFocusedComponent(): JComponent = staticImportsTextArea
 
     val panel: JPanel
-//    private val myUserNameText = JBTextField()
+    private val hardCoreImageBackground = JBTextField()
     private val staticImportsTextArea = JBTextArea(100, 10)
 //    private val myIdeaUserStatus = JBCheckBox("Do you use IntelliJ IDEA? ")
 
-//    var userNameText: String?
-//        get() = myUserNameText.text
-//        set(newText) {
-//            myUserNameText.text = newText
-//        }
+    var hardCoreImageBackgroundPath: String
+        get() = hardCoreImageBackground.text
+        set(newText) {
+            hardCoreImageBackground.text = newText
+        }
 //
 //    var ideaUserStatus: Boolean
 //        get() = myIdeaUserStatus.isSelected
@@ -37,7 +37,7 @@ class AppSettingsComponent {
 
     init {
         panel = FormBuilder.createFormBuilder()
-//            .addLabeledComponent(JBLabel("Enter user nameX: "), myUserNameText, 1, false)
+            .addLabeledComponent(JBLabel("Path to use for background when entering hard-core mode: "), hardCoreImageBackground, 1, false)
 //            .addComponent(myIdeaUserStatus, 1)
             .addLabeledComponentFillVertically("Methods or constants to auto-statically import:", staticImportsTextArea)
             .panel